bug: 셀 업데이트 및 제약조건 버그 수정

+++ Bookjam/API/Entity/PlaceIdActivities.swift
@@ -22,7 +22,7 @@ struct Activities: Codable{
     let info: String
     let capacity: Int?
     let headcount: Int?
-    let totalRating: Int?
+    let totalRating: Float?
     let reviewCount: Int?
     let imageUrl: String?
 //    let like_count: Int?
